backend/requirements.txt
Flask
pymongo
google-generativeai
pyttsx3